chore: 기본 설정
