+++ src/components/layouts/Container/index.tsx
@@ -1,0 +1,5 @@
+const Container = () => {
+	return <div></div>;
+};
+
+export default Container;
